Update style.css
There was a couple bugs with Safari I had to fix. The logo was not displaying and each element the gradient was to be shown on needed to be explicitly declared. I also had to restructure the navigation using grid instead of Flexbox, because the list items would crowd up next to one another when viewed in Safari.